Develop

+++ src/main/java/org/academy/kata/implementation/Mihailll333/FiveImpl.java
@@ -18,7 +18,7 @@ public class FiveImpl implements Five {
             isPrime[i] = true;
         }
 
-        for (int i = 2; i * i <= n; i++) {
+        for (int i = 2; (long) i * i <= n; i++) {
             if (isPrime[i]) {
                 for (int j = i * i; j <= n; j += i) {
                     isPrime[j] = false;
